refacto(cards): refactor get main folio cards workflow
