+++ tests/functional/controllers/main_folios_controller.spec.ts
@@ -1,7 +1,16 @@
 import { test } from '@japa/runner'
 import testUtils from '@adonisjs/core/services/test_utils'
 import sinon from 'sinon'
-import AuthServiceMock from '#tests/mocks/auth_service_mock'
+import { FolioFactory } from '#database/factories/folio'
+import AuthServiceMock, { TEST_AUTH_USER_ID } from '#tests/mocks/auth_service_mock'
+import { CardFactory } from '#database/factories/card'
+import { CardFolioFactory } from '#database/factories/card_folio'
+import { ArtistFactory } from '#database/factories/artist'
+import { RarityFactory } from '#database/factories/rarity'
+import { LegalityFactory } from '#database/factories/legality'
+import { SetFactory } from '#database/factories/set'
+import { SubtypeFactory } from '#database/factories/subtype'
+import { TypeFactory } from '#database/factories/type'
 
 test.group('Main folio controller', (group) => {
   let wardenApiClientStub: sinon.SinonStub
@@ -14,5 +23,562 @@ test.group('Main folio controller', (group) => {
 
   group.teardown(() => {
     wardenApiClientStub.restore()
+  })
+
+  test('mainFolioCards - should return paginated cards from user main folio', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const cards = await CardFactory.createMany(3)
+
+    await CardFolioFactory.merge({
+      cardId: cards[0].id,
+      folioId: mainFolio.id,
+      occurrence: 2,
+    }).create()
+
+    await CardFolioFactory.merge({
+      cardId: cards[1].id,
+      folioId: mainFolio.id,
+      occurrence: 1,
+    }).create()
+
+    await CardFolioFactory.merge({
+      cardId: cards[2].id,
+      folioId: mainFolio.id,
+      occurrence: 3,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.properties(body, ['meta', 'data'])
+    assert.properties(body.meta, ['total', 'currentPage', 'perPage'])
+    assert.equal(body.meta.total, 3)
+    assert.equal(body.meta.currentPage, 1)
+
+    assert.isArray(body.data)
+    assert.equal(body.data.length, 3)
+
+    const firstCardFolio = body.data[0]
+    assert.properties(firstCardFolio, ['id', 'occurrence', 'card'])
+    assert.isString(firstCardFolio.id)
+    assert.isNumber(firstCardFolio.occurrence)
+
+    assert.properties(firstCardFolio.card, ['id', 'imageSmall'])
+    assert.isString(firstCardFolio.card.id)
+    assert.isString(firstCardFolio.card.imageSmall)
+
+    const cardIds = body.data.map((cardFolio: any) => cardFolio.card.id)
+    assert.includeMembers(cardIds, [cards[0].id, cards[1].id, cards[2].id])
+
+    const cardFolioWithOccurrence2 = body.data.find((cf: any) => cf.card.id === cards[0].id)
+    const cardFolioWithOccurrence1 = body.data.find((cf: any) => cf.card.id === cards[1].id)
+    const cardFolioWithOccurrence3 = body.data.find((cf: any) => cf.card.id === cards[2].id)
+
+    assert.equal(cardFolioWithOccurrence2.occurrence, 2)
+    assert.equal(cardFolioWithOccurrence1.occurrence, 1)
+    assert.equal(cardFolioWithOccurrence3.occurrence, 3)
+  })
+
+  test('mainFolioCards - should return empty result when user has no cards in main folio', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    await CardFactory.createMany(3)
+
+    await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 0)
+    assert.equal(body.meta.currentPage, 1)
+    assert.isArray(body.data)
+    assert.equal(body.data.length, 0)
+  })
+
+  test('mainFolioCards - should handle pagination correctly', async ({ client, assert }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    await CardFactory.with('cardFolios', 1, (cardFolios) =>
+      cardFolios.merge([
+        {
+          folioId: mainFolio.id,
+          occurrence: 1,
+        },
+      ])
+    ).createMany(15)
+
+    const page1Response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    page1Response.assertStatus(200)
+    const page1Body = page1Response.body()
+    assert.equal(page1Body.data.length, 10)
+    assert.equal(page1Body.meta.currentPage, 1)
+
+    const firstCardFolio = page1Body.data[0]
+    assert.properties(firstCardFolio, ['id', 'occurrence', 'card'])
+    assert.properties(firstCardFolio.card, ['id', 'imageSmall'])
+
+    const page2Response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 2, limit: 10 })
+
+    page2Response.assertStatus(200)
+    const page2Body = page2Response.body()
+    assert.equal(page2Body.data.length, 5)
+    assert.equal(page2Body.meta.currentPage, 2)
+  })
+
+  test('mainFolioCards - should only return cards from main folio, not secondary folios', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const secondaryFolio = await FolioFactory.merge({
+      userId,
+      isRoot: false,
+    }).create()
+
+    const cards = await CardFactory.createMany(3)
+
+    await CardFolioFactory.merge({
+      cardId: cards[0].id,
+      folioId: mainFolio.id,
+      occurrence: 1,
+    }).create()
+
+    // Add cards to secondary folio (should not be returned in main folio cards)
+    await CardFolioFactory.merge({
+      cardId: cards[1].id,
+      folioId: secondaryFolio.id,
+      occurrence: 2,
+    }).create()
+
+    await CardFolioFactory.merge({
+      cardId: cards[2].id,
+      folioId: secondaryFolio.id,
+      occurrence: 1,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.properties(returnedCardFolio, ['id', 'occurrence', 'card'])
+    assert.equal(returnedCardFolio.card.id, cards[0].id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should validate query parameters', async ({ client }) => {
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 'invalid', limit: 'invalid' })
+
+    response.assertStatus(422)
+    response.assertBodyContains({
+      code: 'E_VALIDATION_ERROR',
+    })
+  })
+
+  test('mainFolioCards - should return 404 when user has no main folio', async ({ client }) => {
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(404)
+    response.assertBodyContains({
+      code: 'E_ROW_NOT_FOUND',
+    })
+  })
+
+  test('mainFolioCards - should require authentication', async ({ client }) => {
+    const response = await client.get('/api/v1/folios/cards').qs({ page: 1, limit: 10 })
+
+    response.assertStatus(401)
+  })
+
+  test('mainFolioCards - should filter cards by name when name parameter is provided', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const pikachuCard = await CardFactory.merge({
+      name: 'Pikachu',
+    }).create()
+
+    const charizardCard = await CardFactory.merge({
+      name: 'Charizard',
+    }).create()
+
+    const bulbasaurCard = await CardFactory.merge({
+      name: 'Bulbasaur',
+    }).create()
+
+    await CardFolioFactory.merge([
+      { cardId: pikachuCard.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: charizardCard.id, folioId: mainFolio.id, occurrence: 2 },
+      { cardId: bulbasaurCard.id, folioId: mainFolio.id, occurrence: 1 },
+    ]).createMany(3)
+
+    const response = await client
+      .get('/api/v1/folios/cards')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10, name: 'Pika' })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, pikachuCard.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should filter cards by rarities when rarities parameter is provided', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const commonRarity = await RarityFactory.merge({ id: 1, label: 'Common' }).create()
+    const rareRarity = await RarityFactory.merge({ id: 2, label: 'Rare' }).create()
+    const artist = await ArtistFactory.create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const commonCard = await CardFactory.merge({
+      name: 'Common Card',
+      rarityId: commonRarity.id,
+      artistId: artist.id,
+    }).create()
+
+    const rareCard = await CardFactory.merge({
+      name: 'Rare Card',
+      rarityId: rareRarity.id,
+      artistId: artist.id,
+    }).create()
+
+    await CardFolioFactory.merge([
+      { cardId: commonCard.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: rareCard.id, folioId: mainFolio.id, occurrence: 2 },
+    ]).createMany(2)
+
+    const response = await client
+      .get(`/api/v1/folios/cards?rarities[]=${commonRarity.id}`)
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, commonCard.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should filter cards by artists when artists parameter is provided', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const artist1 = await ArtistFactory.merge({ id: 1, name: 'Artist One' }).create()
+    const artist2 = await ArtistFactory.merge({ id: 2, name: 'Artist Two' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const card1 = await CardFactory.merge({
+      name: 'Card by Artist One',
+      artistId: artist1.id,
+    }).create()
+
+    const card2 = await CardFactory.merge({
+      name: 'Card by Artist Two',
+      artistId: artist2.id,
+    }).create()
+
+    await CardFolioFactory.merge([
+      { cardId: card1.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: card2.id, folioId: mainFolio.id, occurrence: 2 },
+    ]).createMany(2)
+
+    const response = await client
+      .get(`/api/v1/folios/cards?artists[]=${artist1.id}`)
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, card1.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should filter cards by subtypes when subtypes parameter is provided', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const basicSubtype = await SubtypeFactory.merge({ id: 1, label: 'Basic' }).create()
+    const stage1Subtype = await SubtypeFactory.merge({ id: 2, label: 'Stage 1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const basicCard = await CardFactory.merge({
+      name: 'Basic Card',
+    })
+      .with('subtypes', 1, (subtypes) => subtypes.merge([basicSubtype]))
+      .create()
+
+    const stage1Card = await CardFactory.merge({
+      name: 'Stage 1 Card',
+    })
+      .with('subtypes', 1, (subtypes) => subtypes.merge([stage1Subtype]))
+      .create()
+
+    await CardFolioFactory.merge([
+      { cardId: basicCard.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: stage1Card.id, folioId: mainFolio.id, occurrence: 2 },
+    ]).createMany(2)
+
+    const response = await client
+      .get(`/api/v1/folios/cards?subtypes[]=${basicSubtype.id}`)
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, basicCard.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should filter cards by types when types parameter is provided', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const psychicType = await TypeFactory.merge({ id: 1, label: 'Psychic' }).create()
+    const fireType = await TypeFactory.merge({ id: 2, label: 'Fire' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const psychicCard = await CardFactory.merge({
+      name: 'Psychic Card',
+    })
+      .with('types', 1, (types) => types.merge([psychicType]))
+      .create()
+
+    const fireCard = await CardFactory.merge({
+      name: 'Fire Card',
+    })
+      .with('types', 1, (types) => types.merge([fireType]))
+      .create()
+
+    await CardFolioFactory.merge([
+      { cardId: psychicCard.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: fireCard.id, folioId: mainFolio.id, occurrence: 2 },
+    ]).createMany(2)
+
+    const response = await client
+      .get(`/api/v1/folios/cards?types[]=${psychicType.id}`)
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({ page: 1, limit: 10 })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, psychicCard.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('mainFolioCards - should apply multiple filters simultaneously', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const commonRarity = await RarityFactory.merge({ id: 1, label: 'Common' }).create()
+    const rareRarity = await RarityFactory.merge({ id: 2, label: 'Rare' }).create()
+    const artist1 = await ArtistFactory.merge({ id: 1, name: 'Artist One' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const matchingCard = await CardFactory.merge({
+      name: 'Pikachu Common',
+      rarityId: commonRarity.id,
+      artistId: artist1.id,
+    }).create()
+
+    const nonMatchingCard1 = await CardFactory.merge({
+      name: 'Charizard Common',
+      rarityId: commonRarity.id,
+      artistId: artist1.id,
+    }).create()
+
+    const nonMatchingCard2 = await CardFactory.merge({
+      name: 'Pikachu Rare',
+      rarityId: rareRarity.id,
+      artistId: artist1.id,
+    }).create()
+
+    await CardFolioFactory.merge([
+      { cardId: matchingCard.id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: nonMatchingCard1.id, folioId: mainFolio.id, occurrence: 2 },
+      { cardId: nonMatchingCard2.id, folioId: mainFolio.id, occurrence: 3 },
+    ]).createMany(3)
+
+    const response = await client
+      .get(`/api/v1/folios/cards?rarities[]=${commonRarity.id}`)
+      .header('Authorization', 'Bearer fake-token-for-testing')
+      .qs({
+        page: 1,
+        limit: 10,
+        name: 'Pika',
+      })
+
+    response.assertStatus(200)
+
+    const body = response.body()
+    assert.equal(body.meta.total, 1)
+    assert.equal(body.data.length, 1)
+
+    const returnedCardFolio = body.data[0]
+    assert.equal(returnedCardFolio.card.id, matchingCard.id)
+    assert.equal(returnedCardFolio.occurrence, 1)
   })
 })
